🐞fix: Output value mapping for multiple same records
When dealing with inserts that have the same records inserted several times in
one batch or with updates that update the same records DB generated values did not map correctly.

Reworked the mapping to map updated records by Key & ExplictKey fields. Covered the case where the same recrod could be updated multiple time in the same batch.

Reworked the mapping for insert to not depend on order of output (as it is not guaranteed), rather map by matcing on the available properties. The rationale is that if two records have the same properties except the DB generated ones it does not matter to which object the values are mapped as long as the all values are mapped and each object & output values are mapped only once.