🐞fix: Removed batched bulk source in Execute
BulkExtensions.Execute created batches from data in an attempt to speed up the
bulk operations. In the context of the operations this does not make sense since
only 1 batch was ever created.

The batching code was moved to TransferBulkAsync and Execute was changed to return only a singe source as if there was only one batch.