🐞fix: Mapping return values for enum, datetime
When mapping return values for objects containing enums and datetime the mapping
failed.

Enum value mapping fails as they are stored by their value, so when
matching the results the equality would fail. This has been replaced with enum
specific matcher by value.

DateTime matching fails for two reasons. First if using SqlServer datetime then the precision between .Net and SqlServer is not the same and will fail the match. Second reason is that the SqlServer might be on a different TZ than the server which when using datetime in DB will result failing matches. This has not been resolved properly the datetime match now always returns true.