✨feat: Auto filters & update objects (#7)
* ✨feat: Change tracking proxy factory

Added support for creating change tracking proxies for intrafaces and classes. Proxy tracks changes on writable virtual properties (for classes) or all writable properties (for interfaces).

Changes are availabe through the IChangeTrack interface on the object.

* ✨feat: Automapping of filters & updates

Introduced default auto mappers for entites which provide out-of-the box ability to use column based filtering and updateing. This reduces boilerplate code required to setup an entity. Extended filter and update properties are still supported and full control of the mapping is supported.

BREAKING CHANGE: Removed IQueryFilter interface

* version bump

* Alpha version

* Composite interface builder

* Fix issue with missing properties

* Version fixes & bump

* Final version bump